+++ src/Organizations/Notes/const.js
@@ -1,0 +1,11 @@
+export const ORG_DOMAIN = 'organizations';
+export const ORG_NOTE_TYPE = 'organization';
+export const PANE_HEADER_APP_ICON = 'organizations';
+
+export const ENTITY_TYPE_TRANSLATION_KEYS = {
+  [ORG_NOTE_TYPE]: 'ui-organizations.notes.entity',
+};
+
+export const ENTITY_TYPE_PLURALIZED_TRANSLATION_KEYS = {
+  [ORG_NOTE_TYPE]: 'ui-organizations.notes.entityPlural',
+};