+++ src/Organizations/Organizations.js
@@ -4,6 +4,15 @@ import {
   Route,
 } from 'react-router-dom';
 
+import {
+  NOTES_ROUTE,
+  ORGANIZATIONS_ROUTE,
+} from '../common/constants';
+import {
+  NoteCreate,
+  NoteEdit,
+  NoteView,
+} from './Notes';
 import { OrganizationsListContainer } from './OrganizationsList';
 import { OrganizationCreate } from './OrganizationCreate';
 import { OrganizationEdit } from './OrganizationEdit';
@@ -11,6 +20,21 @@ import { OrganizationEdit } from './OrganizationEdit';
 const Organizations = () => {
   return (
     <Switch>
+      <Route
+        exact
+        path={`${NOTES_ROUTE}/new`}
+        component={NoteCreate}
+      />
+      <Route
+        exact
+        path={`${NOTES_ROUTE}/:id`}
+        component={NoteView}
+      />
+      <Route
+        exact
+        path={`${NOTES_ROUTE}/:id/edit`}
+        component={NoteEdit}
+      />
       <Route
         path="/organizations/create"
         component={OrganizationCreate}
@@ -22,7 +46,7 @@ const Organizations = () => {
       />
 
       <Route
-        path="/organizations"
+        path={ORGANIZATIONS_ROUTE}
         component={OrganizationsListContainer}
       />
     </Switch>