+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Add "add alternative names" button label. Refs UIORGS-205
 * Select list default values not saving in form. Refs UIORGS-203
 * Update view pane to display null as hyphen. Refs UIORGS-201
 * Fix Typing into drop down box, does not pull the type ahead pick-list. Refs UIORGS-198